ocr_service/requirements.txt
# PaddleOCR Service - Isolated Dependencies
# Compatible versions for PaddleOCR 2.7.3

# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.20

# PaddleOCR and dependencies
paddlepaddle==2.6.2
paddleocr==2.7.3

# Image processing - MUST use versions compatible with paddleocr
pillow>=10.0.0
opencv-python<=4.6.0.66  # Required by paddleocr 2.7.3

# Required by PaddleOCR
numpy<2.0.0
pyclipper>=1.3.0
shapely>=2.0.0

# Logging
loguru==0.7.2
